Add adapters for pulumi.Output values.
Add three adapters that allow pulumi.Output values to be used as inputs
to CDK resources:

- asString, which converts an Output to a string
- asNumber, which converts an Output to a number
- asList, which converts an Output to a string[]

Internally, this works by assigning each Output a unique numerical
identifier and creating an intrinsic Token with the value
`{ Ref: { PulumiOutput: <output id> } }`. The use of `Ref` is required
by the CFN validator, which otherwise errors if e.g. a value like
`{ "Pulumi::Output": <output id> }` is passed where a string is
expected.

There is likely more work to be done in the CFN -> Pulumi translator to
handle outputs embedded in strings, function calls, etc.

These changes also implement support for `DependsOn` in order to ensure
that the appsvc test passes reliably.