Add attribute support for CdkResource.
Add a new parameter to the CdkResource constructor that accepts a list
of output attribute names for the resource. This allows the Node SDK to
populate the resource's output attributes. In the CDK adapter, populate
this list of attributes for a resource that represents a CFN element
by scraping unresolved tokens that target the element off of the
element's properties and converting them back into attribute names.

Contributes to #6.