Partially drive conversion from synthesis output.
These changes refactor the conversion process to be partially driven by
the cloud assembly produced by `App.synth`. The assembly is used to
discover the stacks and assets that need to be converted and their
interdependencies. The actual conversion of each stack is still based on
the CFN fragments for each CfnElement, though this will likely change in
the future to be drive from the stack's synthesized template.