Support static file assets.
Static file assets--that is, file assets that are not generated by
running a command--are translated to S3 bucket object resources. As part
of these changes, a new node is created in the resource tree that
corresponds to the CDK stack itself in order to simplify the dependency
graph. This node depends on any assets required by the stack (rather
than plumbing the assets through to each of the stack's children).