Draft: adjust public API
The proposed public API requires that users subclass pulumicdk.Stack
(which is itself a subclass of cdk.Stack) and call `this.synth()` at the
end of the subclass's constructor. `synth` is provided by
`pulumicdk.Stack` and does the work of creating a Pulumi component
resource, synthesizing the CDK resources, etc.