fix!: correctly normalize json types
Some properties in aws-native are Json types, meaning that their values
are sent to CloudControl as is. For all other types we do some
conversion between the CloudControl type and the pulumi type.

An example of this is any policy documents, such as the
`AWS::S3::AccessPoint.Policy` property. The pulumi type is
`pulumi.json#/Any`. In the below example, you can see the property keys
within `policy` have their first character uppercase.

```ts
new aws.s3.AccessPoint('access-point', {
  policy: {
    Version: '2012-10-17',
    Statement: [
      {
        Action: ['s3:GetObject'],
        ...
      }
    ]
  }
});
```

In order to handle these types in the `normalize` function I've added a
utility to parse the aws-native `metadata.json` schema and lookup the
types.

I've also removed the `cfn` mappings that exist in order to handle the
json mappings and the existing tests cover it.

BREAKING CHANGE: `normalize` now takes an additional `cfnType` argument

closes #127