Refactor to use cloud assembly
Previously this library walked the construct tree in-memory and called the
internal `_toCloudFormation()` method each `Cfn*` construct has to get
the CloudFormation resource fragment. This is both inefficient (because
we call `synth` which internally does the same thing, and then we do it
again), and potentially inaccurate (`synth` internally does a lot more
than we do when we walk the construct tree).

This PR is a refactor (maybe closer to a rewrite) that switches to
taking the output of `app.synth()` and processing the resulting
CloudAssembly. The CloudAssembly has everything necessary to do the
conversion and contains the fully resolved CloudFormation template (we don't
have to worry about unresolved tokens!).

This also sets us up for future work like nested stacks, multiple
stacks, and #153.

Note to reviewers. This is a pretty big refactor so I would recommend
reviewing this instead as a restart rather than trying to figure out
what the old code is doing.

Closes #18