Use custom pulumi app synthesizer
Each CDK stack has a [synthesizer](https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib-readme.html#stack-synthesizers)
which determines how the CDK stack should be synthesized and deployed.
By default stacks will use the `DefaultStackSynthesizer` which expects
that the AWS account has been
[bootstrapped](https://docs.aws.amazon.com/cdk/v2/guide/bootstrapping-env.html).
This default bootstrapping will create a bunch of IAM roles, an S3
Bucket for file assets, and an ECR Repository for image assets.

In our case we do not want the user to have to worry about bootstrapping
the account with a CloudFormation template or with the CDK CLI.

This PR creates a new synthesizer (`PulumiSynthesizer`) which will
create the required S3 Bucket and ECR Repository on-demand as needed.

Also, since the synthesizer is also responsible for registering assets
and writing those assets to the asset manfests, our synthesizer is able
to simplify the asset publishing. Instead of registering the assets and
writing the manifest and then post processing that manifest to create
the `BucketObjectV2` resources for each asset in the manifest, we are
able to create those resource as the assets are registered.

closes #108