chore: cdk deps need to be the same for examples
CDK libraries have a weird requirement that only a single version of
`aws-cdk-lib` and `constructs` be installed. This means that the
`devDependency` in `package.json` and the dependencies in each
`examples` `package.json` needs to be the same. If they are different
versions then the examples will fail with `Argument of type 'this' is not assignable to parameter of type 'Construct'.`

fixes #175