Add apigateay integration tests
This adds a new folder for integration tests that are separate from the
`examples`. This will be for tests that should not be used as examples,
but are needed just for test coverage.

As part of this testing I found some issues that needed to be fixed

1. Schema type with `additionalProperties`
Some resources properties are an `object` type with
`additionalProperties`. These properties Should _not_ have their keys
transformed, but should have their values.

2. Returning multiple resources from the mapping function
Sometimes a single CFN resource requires multiple aws classic resources.
In this case we need to return a list of resources with the unique
logicalId of the resources, otherwise each resource gets the same
logicalId and can cause issues with mapping attributes.