Fix resource mappings to multiple resources (#191)
Typically there will be a couple of scenarios for how a CFN resource
maps to Pulumi resources:

1. The CFN resource maps to a single Pulumi aws-native resource
This is the most straightforward case because everything maps directly,
including the attributes

2. The CFN resource maps to a single Pulumi aws resources
In this case the mapping is to a single resource, but some times the
attributes available
on the CFN resource do not map to the attributes on the Pulumi resource.
In these cases
the mapping can return custom attributes that are later available when
references are resolved

3. The CFN resource maps to multiple Pulumi aws resources
One example would be the AWS::IAM::Policy resource which in CFN includes
Role, Group, and User
Polices, but in Pulumi aws these are broken out into separate resources.
In that case the \"main\"
resource would be the Policy and the other resources would be added to
the resources array.
The critical point here is that the \"main\" resource needs to have a
logicalId that matches the
logicalId of the CFN resource, while the other supporting resources must
have different logicalIds.

This PR fixes the 3rd case. Previously we were not handling the other
supporting resources.