Fix IPV6 cidr blocks
Depends on https://github.com/pulumi/pulumi-aws-native/pull/1797

Because of https://github.com/pulumi/pulumi-aws-native/issues/1798 we
need to add a special case for Ipv6 cidr blocks that are added to the
VPC via a `VPCCidrBlock` resource.

This PR adds special handling where we track both the `Vpc` and the
`VpcCidrBlock` resource and swap any references.