Remove CDK generated hash suffix on resource names (#215)
CDK automatically adds a hash suffix to the resource logicalId (which we
then use as the Pulumi Resource name). When pulumi autonaming takes
place we then append an additional random suffix so we end up with names
like `testapimyapiCloudWatchRoleA84BAE8E-120b18f`. Apart from this
looking bad it also takes up 16 characters for each name and increases
the likelihood of running into naming limits.

This PR overrides the `allocateLogicalId` method on the `cdk.Stack`.
Most of the code is simply forked from the CDK implementation with the
removal of the piece that generates the hash.

You can view the CDK version here

https://github.com/aws/aws-cdk/blob/ccab485b87a7090ddf0773508d7b8ee84ff654b0/packages/aws-cdk-lib/core/lib/stack.ts?plain=1#L1357

closes #185