CDK RemovalPolicy maps to Pulumi retainOnDelete
This PR maps CloudFormation [DeletionPolicy](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-attribute-deletionpolicy.html)
to Pulumi `retainOnDelete`.

**`DeletionPolicy` => `retainOnDelete`**

- `Delete` => `false`
- `Retain` => `true`
- `RetainExceptOnCreate` => `true`
- `Snapshot` => `true`

Looking specifically for validation on `Snapshot` mapping. This
functionality does not exist in Pulumi so I decided to err on the side
of caution and retain the resource instead of deleting it. We log a
warning in this case and they should see the retention in the preview.

closes #188