Handle pulumi.Output recursively when translating CDK properties to Pulumi (#228)
`pulumi.Output` values seem to be able to reach the `normalize` step of
translating properties from a CDK notation to a Pulumi notation and
making sure the casing is right. This may be happening in particular
because CF intrinsic evaluator may emit Output values. With the fix, the
`normalize` step can deeply recur into these eventual values and make
sure that the property casing is correctly handled regardless of whether
a value is an eventual or not.

This seems to resolve an issue I have been running into when testing a
Kinesis Stream.