Validate all CDK resources are mapped to Pulumi resources
Previously we didn't ensure that all CDK resources were actually
mapped to Pulumi resources.
While working on adding support for Custom Resources we noticed
that some resources were not mapped (e.g. resources with ID of 'Default').

This change adds validation to the GraphBuilder in order to ensure
that all resources are mapped.
In order to prevent any unexpected behavior we're failing hard if we do
not detect that all resources are mapped.