chore: add example of looking up availability zones
If you provide an explicit environment to the stack (there are a lot of
reasons to do this) then you have to enable lookups in order for the
stack to lookup the availability zones.

It would be nice if you could just use Pulumi functions to lookup the
AZs and provide that, but the values have to be strings (can't be
promises/outputs). An easy workaround for this is to use the `getAzs`
intrinsic to defer the lookup and allow the intrinsic to be converted to
a Pulumi function.

I am adding an example of how to do this. The reason we don't want to do
this by default is that the number of availability zones is not
something that should every change without the user knowing about it.
CDK handles that by having the lookup store the value in the
`cdk.context.json` file so the lookup is never re-run.