Improve the Ref intrinsic implementation (#243)
This change adds some tests and comments around the `Ref` intrinsic
implementation and moves it to `intrinsics.ts`. The one change in
behavior is that now `Ref` is able to consult Pulumi metadata from the
`aws-native` provider. The metadata helps it resolve to the same values
as it would in CloudFormation for resources that have non-standard `Ref`
behavior. The correctness of this process is up to the correctness of
the metadata tables. Entries not found in the metadata tables default to
using Pulumi `id` for most cases should do the right thing.

Fixes https://github.com/pulumi/pulumi-cdk/issues/173

Needs https://github.com/pulumi/pulumi-aws-native/pull/1836

---------

Co-authored-by: Cory Hall <43035978+corymhall@users.noreply.github.com>