Reference assets directly
Previously we were generating a physical name for the staging resources
(S3 Bucket / ECR Repository) because it was mistakenly assumed that we
needed those to be known values for CDK to work. Turns out that we can
use `asString` on the Output values and CDK handles it correctly. So
this PR does a couple of things:

1. Removes the hard coded resource names and allows Pulumi to generate
   an autoname.
2. Removes the dependency each stack had on the staging stack. Because
   we are using Output values to reference the assets, each resource can
   depend on the individual asset directly. This allows for more
   resources to be provisioned in parallel. On the `eventbridge-sns`
   test this had a 20% decrease in deployment time.