Do not use Ref metadata for AWS Provider resources
We use the `Ref` metadata to determine which attribute the `Ref`
intrinsic should reference. This metadata is only valid for CCAPI
resources though so we need to exclude any other resource type.

This PR checks the internal `__pulumiType` to determine if the resource
is an `aws-native` resource. I chose to do it this way because we need
to check the resource itself, we can't rely on any other information
source. For example, the user could provide their own mapping function
that maps some resources to `aws-native` and some resources to `aws`.

closes #262