Add a postUpgradeTask to bump metadata for aws-native (#267)
This adds a
[postUpgradeTask](https://github.com/pulumi/renovate-config?tab=readme-ov-file#post-upgrade-tasks)
to update metadata whenever `@pulumi/aws-native` is bumped.

The metadata bump is straightforward enough that we should be able to
run it inside Renovate without trouble.

This is defined as a make target because Renovate requires post-run
tasks to be explicitly allowed, and for the moment we allow `make
renovate` for simplicity. We can think about expanding allowable
commands in the future.