Start renovating example and integration dependencies (#273)
We currently allow example and integration dependencies to float, which
means PRs can begin to fail consistently if an upstream dependency
breaks.

We should be able to detect these breakages quickly, and we do now.
https://github.com/pulumi/pulumi-cdk/pull/272 failed within an ~hour of
the AWS release.

However we also don't want upstream issues to block us. This project is
currently unable to release due to ongoing issues with p-aws, for
example.

This PR does two things:
* It pins our p-aws dependency to known-good v6.61.0 across `./examples`
and `./integration`. This should allow tests to start passing while the
p-aws mitigation continues.
* It begins committing `yarn.lock` files, which will enable Renovate to
start maintaining dependencies for these directories.

The net result will be:
* More hermetic and consistent versioning of dependencies across the
provider and tests. Re-running a PR's tests after a week won't run them
against a different version of pulumi/pulumi, for example.
* PR tests and upstream dependencies are decoupled, so upstream p0s
shouldn't impact us until we attempt to upgrade the dependency.