Enhance CDK support with nested stack handling (#295)
This change adds support for Nested Stacks to `pulumi-cdk`.
In CDK, a nested stack is a CDK stack that you create inside another
stack, known as the parent stack. Users create nested stacks using the
[NestedStack](https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.NestedStack.html)
construct.
By using nested stacks, users can organize resources across multiple
stacks. Nested stacks also offer a way around the AWS CloudFormation
500-resource limit for stacks. A nested stack counts as only one
resource in the stack that contains it. However, it can contain up to
500 resources, including additional nested stacks.
While Pulumi doesn't have those limitations, many existing CDK
constructs use nested stacks to get around the CloudFormation
limitations. The EKS cluster construct is one of the most prominent ones
that require nested stacks.

In detail, this change:
- enhances the `AssemblyManifestReader` to support reading nested
stacks.
- introduces a new unique `StackAddress` for elements in the
CloudFormation templates. It combines the stack path (tree path of the
stack) and the stack-local logical ID in order to form an assembly wide
unique address. This is necessary in order to represent a hierarchy of
stacks in the same pulumi program
- adds a new `NestedStackConstruct` component. It is used to namespace
the resources of a NestedStack in order to not run into any issues with
duplicate logical IDs. It achieves this by including the stack path in
the pulumi resource type. The types of parent resources are included in
URNs, meaning all resources parented under the nested stack have their
own unique namespace.
- updates the artifact converter and intrinsics to handle cross stack
references (Parameters & Outputs) for nested stacks

Integration tests and an example are part of
https://github.com/pulumi/pulumi-cdk/pull/298 which is part of this PR
stack.