Fix permanent docker asset diff (#318)
Docker assets show a permanent diff on preview due to the `cdk.out`
directory using an absolute path to a system temporary directory. Even
if the asset does not change, the temporary directory will always be
different.

There are two ways we could go about solving this:

1. Update our internally created `cdk.App` to always set `outdir` to a
relative `cdk.out` directory
2. Update our docs to point users to set this themselves. 

I opted for the second route mainly because this follows the default
behavior of CDK. Almost all CDK applications will have a `cdk.json` file
to configure some basic settings, one of them being the `outdir`
setting. We don't require users to have this file, but we should be
strongly encouraging users to have this because there are a couple of
additional things that users should set (`app`, and `context` feature
flags).

I've updated the readme to have an additional `Getting Started` section
which matches what we have in
https://www.pulumi.com/docs/iac/clouds/aws/guides/cdk/. I've added the
`cdk.json` part (i'll need to also go back and add this to the guide).

closes #287