Automate renovate updates for CDK dependencies (#401)
Follow up to #345 which was supposed to automate the CDK upstream
updates, but did not. The previous PRs only enabled us to manually
trigger the upgrades. This PR sets `dependencyDashboardApproval: false`
which is required for renovate to automatically create the PR when there
are version bumps.

This also adds the `aws-cdk-lib` dependency to the cdk group so that the
cdk dependencies are upgraded together.